requirements.txt
# python
python-dotenv

# data
ibis-framework[duckdb,polars,deltalake,bigquery] @ git+https://github.com/ibis-project/ibis.git@master

# viz
plotly
streamlit

